pyproject.toml
[project]
name = "ai-engineering-fellowship"
version = "0.1.0"
description = "AI Engineering Fellowship"
requires-python = ">=3.11"
dependencies = [
    "langchain-core",
    "langgraph",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rank-bm25",
    "pypdf",
    "ipykernel",
    "reportlab>=4.4.9",
    "qdrant-client>=1.16.2",
    "networkx>=3.6.1",
    "streamlit>=1.41.0",
]
